+++ src/main/scala/edu/ie3/simona/service/load/LoadProfileStore.scala
@@ -15,8 +15,8 @@ import edu.ie3.simona.model.participant.load.ProfileLoadModel.ProfileLoadFactory
 import edu.ie3.simona.util.SimonaConstants.FIRST_TICK_IN_SIMULATION
 import edu.ie3.simona.util.TickUtil.RichZonedDateTime
 import edu.ie3.util.scala.quantities.QuantityConversionUtils.{
-  EnergyToSimona,
-  PowerConversionSimona,
+  toApparent,
+  toSquants,
 }
 import tech.units.indriya.ComparableQuantity